Add a property scale
The new property `scale` enables user to to change the scale factor of
the frame for face detection and tracking. Especially for 4K, it is
recommended to scale more so that the CPU usage is reduced.
Since dlib requires 80x80 for face detection, user need to adjust the
scale factor for each usage.
This change also added the scale operation in GPU.
Close #32.