Reduce log message if failed to load face-tracker.effect
There is a bug report #48 containing a lot of `effect_setval_inline:
invalid param`. I thinks this is because the failure of loading
`face-tracker.effect`.
To avoid a lot of error messages, `effect_ft` is checked and not to use
the effect if it was failed to load the effect file.