Add a new source Face Tracker Monitor
A new source `face_tracker_monitor` is added to display face detection
and tracking status in another source. It will be useful to monitor the
status while the face-tracker is working for the program.
The status is drawn using newly added proc-handlers on the filter.
Since this source will be a part of the dock, which will be implemented
later, the source is marked OBS_SOURCE_CAP_DISABLED by default.