cmake(windows): Suppress warnings for fopen
To avoid warnings claiming `fopen` is unsafe and suggesting to use
`fopen_s` instead, a macro `_CRT_SECURE_NO_WARNINGS` is defined in
cmake.
I don't want to disable all unsafe warnings but I cannot find a way to
suppress the warnings only for `fopen`.

The difference between `fopen` and `fopen_s` is returning error-code or
not. I think `fopen` itself is not unsafe but reading `errno` is unsafe.